backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
yfinance==0.2.28
textblob==0.17.1
requests==2.31.0
beautifulsoup4==4.12.2
gunicorn==21.2.0